refactor: Organize content messages into year-specific directories and update all related imports.
