+++ app/[locale]/(2026)/layout.tsx
@@ -1,7 +1,7 @@
 import { setRequestLocale } from "next-intl/server";
 import Nav from "./_components/Nav";
 import Footer from "./_components/Footer";
-import navItems from "@/app/messages/nav";
+import navItems from "@/app/messages/2025/nav";
 import ColorBends from "@/components/ColorBends";
 import FloatingLines from "@/components/FloatingLines";
 import LiquidEther from "@/components/LiquidEther";
