+++ app/[locale]/2025/_components/Nav/index.tsx
@@ -4,7 +4,7 @@ import { useCallback, useState } from "react";
 import Image from "next/image";
 import { Link, useRouter } from "@/i18n/navigation";
 import { useLocale } from "next-intl";
-import { type NavItem } from "@/app/messages/nav";
+import { type NavItem } from "@/app/messages/2025/nav";
 import Mobile from "./Mobile";
 
 type Props = {
@@ -69,7 +69,7 @@ export default function Nav({ items }: Props) {
               ))}
               <Link
                 href="/"
-                locale={locale === "en" ? "ko" : "en"}
+                locale={locale === "en" ? "ko/2025" : "en/2025"}
                 className="text-white/80 hover:text-white px-3 py-2 text-sm font-medium transition-colors duration-200 bg-white/10 hover:bg-white/20 rounded-md border border-white/20 cursor-pointer"
               >
                 {locale === "en" ? "한국어" : "English"}
