+++ app/[locale]/2025/_components/Sponsor.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import sponsors, { type Sponsor } from "../../../messages/sponsors";
+import sponsors, { type Sponsor } from "../../../messages/2025/sponsors";
 import ShinyText from "@/components/ShinyText";
 import GradientText from "@/components/GradientText";
 import { cn } from "@/lib/utils";
