+++ app/[locale]/2025/page.tsx
@@ -4,11 +4,11 @@ import Schedule from "./_components/Schedule";
 import Speakers from "./_components/Speakers";
 import Sponsor from "./_components/Sponsor";
 import Divider from "./_components/Divider";
-import speakers from "@/app/messages/speakers";
-import schedules from "@/app/messages/schedules";
-import tickets from "@/app/messages/tickets";
-import attendees from "@/app/messages/attendees";
-import markets from "@/app/messages/markets";
+import speakers from "@/app/messages/2025/speakers";
+import schedules from "@/app/messages/2025/schedules";
+import tickets from "@/app/messages/2025/tickets";
+import attendees from "@/app/messages/2025/attendees";
+import markets from "@/app/messages/2025/markets";
 import Section from "./_components/Section";
 import Attendees from "./_components/Attendees";
 import LightningMarket from "./_components/LightningMarket";
