+++ app/messages/2025/en.json
@@ -1,0 +1,54 @@
+{
+  "Hero": {
+    "title": "Bitcoin Mini Conference",
+    "subtitle": "SECOND EDITION",
+    "date": "2025.11.29-30",
+    "location": "Seoul · South Korea"
+  },
+  "Speakers": {
+    "title": "Speakers",
+    "moreText": "More",
+    "closeText": "Close"
+  },
+  "Schedule": {
+    "title": "Schedule",
+    "triggerText": "View details"
+  },
+  "Video": {
+    "title": "Conference Highlights",
+    "description": "Watch the teaser video of our conference"
+  },
+  "LightningMarket": {
+    "title": "Lightning Market",
+    "moreText": "Details",
+    "about": "About",
+    "products": "Products",
+    "specialEvent": "Special Event",
+    "productsInformationComingSoon": "Products information coming soon...",
+    "clickToViewDetails": "Click to view details",
+    "details": "Details",
+    "onlineStore": "Online Store",
+    "website": "Website",
+    "showMoreText": "More",
+    "showCloseText": "Close"
+  },
+  "Sponsor": {
+    "title": "Sponsors",
+    "gold": "Gold Sponsor",
+    "silver": "Silver Sponsor",
+    "bronze": "Bronze Sponsor",
+    "individual": "Individual Sponsor"
+  },
+  "Footer": {
+    "contact": "Contact",
+    "email": "Email:",
+    "donation": "Donation",
+    "location": "Location",
+    "day1": "Nov 29",
+    "day2": "Nov 30",
+    "address1": "NSP Hall, B1, 20, Magokjungang 1-ro, Gangseo-gu, Seoul, Republic of Korea",
+    "address2": "Hanbit Building, 76, Yeonhui-ro 2-gil, Seodaemun-gu, Seoul, Republic of Korea",
+    "address3": "SPACE BE, 21, Donggyo-ro 30-gil, Mapo-gu, Seoul, Republic of Korea",
+    "copyright": "© 2025 Bitcoin Mini Conference. All rights reserved."
+  }
+}
